Intial commit to deploy the external DNS service
