+++ projects/core/src/coredns.ts
