+++ projects/core/src/index.ts
@@ -1,4 +1,5 @@
-import './dns';
+import './coredns';
+import './externaldns';
 import './tailscale';
 import './elastic';
 import './service.dashboard';
