+++ projects/core/src/traefik/index.ts
@@ -1,22 +1,7 @@
 import * as k8s from '@pulumi/kubernetes';
 import provider from '../provider';
 import { Config } from '@pulumi/pulumi';
-
-const config = new Config();
-const settings = config.requireObject('cloudflare') as {
-  userEmail: string;
-  apiToken: string;
-};
-
-const cloudflareSecret = new k8s.core.v1.Secret('cloudflare-api-token', {
-  metadata: {
-    name: 'cloudflare-api-token',
-  },
-  stringData: {
-    CF_API_EMAIL: process.env[settings.userEmail] + '',
-    CF_DNS_API_TOKEN: process.env[settings.apiToken] + '',
-  },
-});
+import { cloudflareSecret } from '../cloudflare';
 
 export const traefik = new k8s.helm.v3.Chart(
   'traefik',
